requirements.txt
Django>=3.1.2
djangorestframework
psycopg2-binary>=2.8
tensorflow
sklearn
sklearn-crfsuite==0.3.6
pandas
pickleshare==0.7.5
PyMySQL==0.10.1
keras
requests
Pillow